Fix missing parameter rename
In 27d1053a27e2640de13071ec8f4badf8a8f9823e the signature of
save_channel() was updated and the users variable was removed from the
main function but one save_channel() call was still using it.

Fix that by calling save_channel() with user_list in the missing
instance.